Group the Nemotron-H patches into an architectures package
The Liger kernels and the Mamba2 mixer workarounds were split across
llm_ops and third_party_patches with no shared structure. They now sit under
agilerl/architectures/nemotron_h, behind a family registry that
install_zero3_third_party_hooks dispatches through, so a second architecture
is a new subpackage and a table entry.

third_party_patches keeps the generic ZeRO-3 patches. The resolution and
idempotence primitives move down to agilerl/utils/patching so the generic and
per-family modules can share them without importing each other.

No behaviour change: the same patches install at the same points for the same
checkpoints.
