+++ tests/test_architectures/__init__.py
@@ -1,0 +1,2 @@
+# Copyright 2026 AgileRL
+# SPDX-License-Identifier: Apache-2.0
